requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
idna==2.6
python-dotenv==0.7.1
requests==2.18.4
six==1.10.0
slackclient==1.0.9
urllib3==1.22
websocket-client==0.44.0
